refactor: rename QR modules by responsibility
