+++ src/components/qr-display.tsx
@@ -2,7 +2,7 @@ import { useEffect, useRef, useState, type ReactNode } from "react"
 import { AlertCircle, Expand, X } from "lucide-react"
 import { renderQrDataUrl } from "@/qr/encode"
 import type { QrEcLevel } from "@/schemas/domain"
-import { isQrCryptPayload } from "@/qr/payload-v2"
+import { isQrCryptPayload } from "@/qr/wire-codec"
 import { toAppError } from "@/crypto/errors"
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert"
 import { Button } from "@/components/ui/button"
