+++ src/pages/encrypt-page.tsx
@@ -62,7 +62,7 @@ import { copyTextToClipboard } from "@/lib/clipboard"
 import { effectiveGeneratedDisplay } from "@/lib/generated-display"
 import { MAX_SYM_PLAINTEXT_BYTES } from "@/lib/limits"
 import { exportQrFramePayloads } from "@/qr/export-frames"
-import { encodeFrameToPayload } from "@/qr/payload-v2"
+import { encodeFrameToPayload } from "@/qr/wire-codec"
 import { type UiAlgorithm } from "@/schemas/domain"
 import { env } from "@/schemas/env-schema"
 import { qrNameSchema } from "@/schemas/key-schema"
