+++ src/qr/camera-scan.ts
