+++ src/qr/wire-codec.ts
