+++ tests/ui/qr-scanner.test.tsx
@@ -11,7 +11,7 @@ import { Dialog, DialogContent, DialogTitle } from "@/components/ui/dialog"
 import { AppError, messageFor } from "@/crypto/errors"
 import { MultipartScanSession } from "@/features/multipart-scan-session"
 import { translate } from "@/i18n/messages"
-import type { QrScanHandle } from "@/qr/decode"
+import type { QrScanHandle } from "@/qr/camera-scan"
 import type { TransferState } from "@/qr/multipart/transfer-state"
 import { deferred } from "../helpers/deferred"
 import { probeWebAssemblyRuntime } from "./helpers/fakes/feature-detection"
