+++ tests/unit/qr-camera/scanner-harness.ts
@@ -183,11 +183,11 @@ export async function advance(ms: number): Promise<void> {
   await flushMicrotasks()
 }
 
-export async function loadColdDecoder(): Promise<typeof import("@/qr/decode")> {
-  return import("@/qr/decode")
+export async function loadColdDecoder(): Promise<typeof import("@/qr/camera-scan")> {
+  return import("@/qr/camera-scan")
 }
 
-export async function loadDecoder(): Promise<typeof import("@/qr/decode")> {
+export async function loadDecoder(): Promise<typeof import("@/qr/camera-scan")> {
   const decoder = await loadColdDecoder()
   await decoder.warmQrReader()
   // The synthetic warm-up probe is not part of the frame pump's decode cadence.
